Map functions changes, fixed filter, changes in custom marker info window, changed app icon
