+++ Mobile/app/src/main/java/com/example/nurkowapolskaapp/app/functions/map/functions/MapFunctions.kt
@@ -1,0 +1,20 @@
+package com.example.nurkowapolskaapp.app.functions.map.functions
+
+import android.os.Build
+import java.text.SimpleDateFormat
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.util.Calendar
+import java.util.Locale
+
+fun getCurrentDate(): String {
+    return if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+        val currentDateTime = LocalDateTime.now()
+        val formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy")
+        currentDateTime.format(formatter)
+    } else {
+        val calendar = Calendar.getInstance().time
+        val formatter = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
+        formatter.format(calendar)
+    }
+}
